pyproject.toml
[project]
name = "mnemosyne"
version = "0.1.0"
description = "Local-first Obsidian vault assistant with CLI, TUI, and Telegram bot."
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "textual>=0.60",
    "python-telegram-bot>=20.0",
    "ollama>=0.3",
    "qdrant-client>=1.9",
    "python-frontmatter>=1.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pytest",
    "pytest-asyncio"
]

[tool.setuptools.packages.find]
where = ["mnemosyne"]
